leetcode/0106.ConstructBinaryTreefromInorderandPostorderTraversal/106. Construct Binary Tree from Inorder and Postorder Traversal.go
package leetcode

type TreeNode struct {
	Val   int
	Left  *TreeNode
	Right *TreeNode
}

func buildTree(inorder []int, postorder []int) *TreeNode {
	if len(inorder) == 0 && len(postorder) == 0 {
		return nil
	}
	root := &TreeNode{
		Val: postorder[len(postorder)-1],
	}

	index := 0
	for i, v := range inorder {
		if v == root.Val {
			index = i
		}
	}
	root.Left = buildTree(inorder[0:index], postorder[0:index])
	root.Right = buildTree(inorder[index+1:], postorder[index:len(postorder)-1])
	return root
}
